Fully remove old osu_migration checks for password hashing.
