+++ app/Providers/AppServiceProvider.php
@@ -50,5 +50,7 @@ class AppServiceProvider extends ServiceProvider
             'Illuminate\Contracts\Auth\Registrar',
             'App\Services\Registrar'
         );
+
+        $this->app->bind('hash', 'App\Hashing\OsuHasher');
     }
 }
